+++ src/renderer/widgets/RemoveVotesModal/default/components/RemoveVotesDefault.tsx
