+++ src/renderer/widgets/RemoveVotesModal/default/model/removeVotesModal.ts
