+++ src/renderer/widgets/RemoveVotesModal/index.ts
@@ -1,3 +1,2 @@
 export { RemoveVotesModal } from './ui/RemoveVotesModal';
-export { RemoveVotesShardsModal } from './shards/components/RemoveVotesShardsModal';
-export { removeVotesModel } from './model/removeVotesModal';
+export { removeVotesModel } from './default/model/removeVotesModal';
